pyproject.toml
[tool.poetry]
name = "fetch-uagent"
version = "0.1.0"
description = "Template of Temperature Alert Agent"
authors = ["Darshan59063 <90100530+Darshan59063@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
aiohttp = "3.8.5"
aiosignal = "1.3.1"
apispec = "6.3.0"
asttokens = "2.4.0"
async-timeout = "4.0.3"
attrs = "23.1.0"
backcall = "0.2.0"
bech32 = "1.2.0"
blinker = "1.6.2"
build = "0.10.0"
cachecontrol = "0.13.1"
certifi = "2023.7.22"
charset-normalizer = "3.3.0"
cleo = "2.0.1"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.1.4"
cosmpy = "0.9.1"
crashtest = "0.4.1"
debugpy = "1.8.0"
decorator = "5.1.1"
distlib = "0.3.7"
dulwich = "0.21.6"
ecdsa = "0.18.0"
exceptiongroup = "1.1.3"
executing = "2.0.0"
filelock = "3.12.4"
flask = "3.0.0"
flask-sqlalchemy = "3.1.1"
frozenlist = "1.4.0"
googleapis-common-protos = "1.60.0"
greenlet = "3.0.0"
grpcio = "1.51.1"
h11 = "0.14.0"
idna = "3.4"
importlib-metadata = "6.8.0"
installer = "0.7.0"
ipykernel = "6.25.2"
ipython = "8.16.1"
itsdangerous = "2.1.2"
jaraco-classes = "3.3.0"
jedi = "0.19.1"
jinja2 = "3.1.2"
joblib = "1.3.2"
jsonschema = "4.17.3"
jsonschema-specifications = "2023.7.1"
jupyter-client = "8.3.1"
jupyter-core = "5.3.2"
keyring = "24.2.0"
markupsafe = "2.1.3"
matplotlib-inline = "0.1.6"
more-itertools = "10.1.0"
msgpack = "1.0.7"
multidict = "6.0.4"
nest-asyncio = "1.5.8"
numpy = "1.26.0"
packaging = "23.2"
pandas = "2.1.1"
parso = "0.8.3"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
plyer = "2.1.0"
poetry = "1.6.1"
poetry-core = "1.7.0"
poetry-plugin-export = "1.5.0"
prompt-toolkit = "3.0.39"
protobuf = "4.24.3"
psutil = "5.9.5"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pycryptodome = "3.19.0"
pydantic = "1.10.13"
pygments = "2.16.1"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pywin32 = "306"
pywin32-ctypes = "0.2.2"
pyzmq = "25.1.1"
rapidfuzz = "2.15.1"
referencing = "0.30.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rpds-py = "0.10.3"
scikit-learn = "1.3.1"
scipy = "1.11.3"
shellingham = "1.5.3"
six = "1.16.0"
sqlalchemy = "2.0.21"
stack-data = "0.6.3"
threadpoolctl = "3.2.0"
tomli = "2.0.1"
tomlkit = "0.12.1"
tornado = "6.3.3"
traitlets = "5.10.1"
trove-classifiers = "2023.9.19"
typing-extensions = "4.8.0"
tzdata = "2023.3"
uagents = "0.6.2"
urllib3 = "2.0.6"
uvicorn = "0.20.0"
virtualenv = "20.24.5"
wcwidth = "0.2.8"
websockets = "10.4"
werkzeug = "3.0.0"
yarl = "1.9.2"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
